+++ src/AbyssIrc.Server/Interfaces/Services/System/ISessionManagerService.cs
@@ -7,7 +7,7 @@ public interface ISessionManagerService
 {
     void AddSession(string id, string ipEndpoint, IrcSession? session = null);
 
-    IrcSession GetSession(string id);
+    IrcSession? GetSession(string id);
 
     List<IrcSession> GetSessions();
 }