+++ src/AbyssIrc.Server/Modules/Scripts/LoggerModule.cs
@@ -17,6 +17,30 @@ public class LoggerModule
     [ScriptFunction("Log an informational message")]
     public void Info(string message, params object[]? args)
     {
-        _logger.LogInformation(message, args);
+        _logger.LogInformation("[JS] " + message, args);
+    }
+
+    [ScriptFunction("Log a warning message")]
+    public void Warn(string message, params object[]? args)
+    {
+        _logger.LogWarning("[JS] " + message, args);
+    }
+
+    [ScriptFunction("Log an error message")]
+    public void Error(string message, params object[]? args)
+    {
+        _logger.LogError("[JS] " + message, args);
+    }
+
+    [ScriptFunction("Log a critical message")]
+    public void Critical(string message, params object[]? args)
+    {
+        _logger.LogCritical("[JS] " + message, args);
+    }
+
+    [ScriptFunction("Log a debug message")]
+    public void Debug(string message, params object[]? args)
+    {
+        _logger.LogDebug("[JS] " + message, args);
     }
 }